# packages/contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: The Penthouse API
  version: 1.0.0-alpha
  description: Private social platform API. All protected endpoints require a Bearer access token.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
              description: Human-readable error message
            - type: object
              description: Zod field-level validation error (fieldErrors + formErrors)
...
    UserRole:
      type: string
      enum: [admin, member]
...
    AuthUser:
      type: object
      required: [id, username, displayName, avatarUrl, role, mustChangePassword]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        mustChangePassword:
          type: boolean
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken]
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
        accessToken:
          type: string
          description: Short-lived JWT (default 15m)
        refreshToken:
          type: string
          description: Long-lived opaque token for rotation (default 7d)
        recoveryCode:
          type: string
          description: One-time recovery code display for signup, reset, or backfilled older accounts
    MeResponse:
      allOf:
        - $ref: '#/components/schemas/AuthUser'
        - type: object
          required: [bio, avatarMediaId]
          properties:
            bio:
              type: string
              nullable: true
            avatarMediaId:
              type: string
              format: uuid
              nullable: true
    UpdateProfileRequest:
      type: object
      minProperties: 1
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 40
        bio:
          type: string
          maxLength: 160
          nullable: true
        avatarUploadId:
          type: string
          format: uuid
          nullable: true
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 10
          maxLength: 128
        newPassword:
          type: string
          minLength: 10
          maxLength: 128
          description: Must not start or end with spaces
    RotateRecoveryCodeResponse:
      type: object
      required: [recoveryCode]
      properties:
        recoveryCode:
          type: string
...
    MemberSummary:
      type: object
      required: [id, username, displayName, avatarUrl]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
    MemberDetail:
      allOf:
        - $ref: '#/components/schemas/MemberSummary'
        - type: object
          required: [bio]
          properties:
            bio:
              type: string
              nullable: true
...
paths:
...
  /api/v1/me:
    get:
      summary: Get current member profile
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current member profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized or invalid user session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/profile:
    patch:
      summary: Update current member profile
      operationId: updateMeProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated member profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '400':
          description: Validation error or invalid avatar selection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Password change required or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/password:
    post:
      summary: Change current member password
      operationId: changeMePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed, recovery code rotated, and fresh session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/recovery-code/rotate:
    post:
      summary: Rotate the current member recovery code
      operationId: rotateMeRecoveryCode
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New recovery code issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotateRecoveryCodeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/members:
    get:
      summary: Search active members
      operationId: listMembers
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: Case-insensitive search across username and display name
          schema:
            type: string
      responses:
        '200':
          description: Active member directory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Password change required or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/members/{memberId}:
    get:
      summary: Get a public-facing active member profile
      operationId: getMemberById
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active member profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Password change required or account unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...